Support for {@param ...}, aka "header params"
Upstream Closure Templates has removed support for params specified in
soydoc and now requires them specified using "header params". This
change adds support for parsing header params and treating them
similarly to soydoc params, having an unknown type.

For backwards compatibility, the header params are added to
`Template.SoyDoc.Params`, so calling code should not require updates.

The result is that this library will not report errors that the Java
library would report, but that level of compatibility is sufficient
for our purposes, and a lot of development would be required to
support the param type system.